Split version related methods into separate file
