+++ govcd/api_vcd_test.go
@@ -628,53 +628,3 @@ func TestVCDClient_Authenticate(t *testing.T) {
 		t.Fatalf("Error authenticating: %v", err)
 	}
 }
-
-func (vcd *TestVCD) Test_APIMaxVerIs(check *C) {
-
-	// vCD 8.20 introduced API version 27.0
-	r := vcd.client.APIMaxVerIs(">= 27.0")
-	check.Assert(r, Equals, true)
-
-	// Mocked tests
-	mockVcd := &VCDClient{
-		SupportedVersions: supportedVersions{
-			VersionInfos{
-				VersionInfo{
-					Version: "27.0",
-				},
-			},
-		},
-	}
-
-	var versionTests = []struct {
-		version      string
-		boolChecker  Checker
-		isSsupported bool
-	}{
-		{"= 27.0", Equals, true},
-		{">= 27.0", Equals, true},
-		{">= 25.0, <= 30", Equals, true},
-		{"> 27.0", Equals, false},
-		{"< 27.0", Equals, false},
-		{"invalid", Equals, false},
-		{"", Equals, false},
-	}
-
-	for _, tt := range versionTests {
-		r := mockVcd.APIMaxVerIs(tt.version)
-		check.Assert(r, tt.boolChecker, tt.isSsupported)
-	}
-}
-
-func (vcd *TestVCD) Test_validateAPIVersion(check *C) {
-	// valid version is checked automatically in SetUpSuite
-	unsupportedVersion := "999.0"
-
-	config, err := GetConfigStruct()
-	check.Assert(err, IsNil)
-
-	vcdClient, err := GetTestVCDFromYaml(config, WithAPIVersion(unsupportedVersion))
-	check.Assert(err, IsNil)
-	err = vcdClient.validateAPIVersion()
-	check.Assert(err, ErrorMatches, "API version .* is not supported: version = .* is not supported")
-}
